[MOVE] <calendar module> set file structure
